add monitor

+++ im/im-domain/src/main/java/com/sparrow/chat/repository/QunRepository.java
@@ -3,7 +3,7 @@ package com.sparrow.chat.repository;
 import java.util.List;
 
 public interface QunRepository {
-    List<Integer> getUserIdList(String qunId);
+    List<Long> getUserIdList(String qunId);
 
-    void syncQunMember(Long qunId,List<Integer> memberIds);
+    void syncQunMember(Long qunId,List<Long> memberIds);
 }
